+++ crates/ironcalc-base/src/expressions/parser/static_analysis.rs
@@ -927,6 +927,13 @@ fn get_function_args_signature(kind: &Function, arg_count: usize) -> Vec<Signatu
             result
         }
         Function::Choosecols | Function::Chooserows => vec![Signature::Vector; arg_count],
+        Function::Tocol | Function::Torow => {
+            let mut result = vec![Signature::Vector; arg_count];
+            for item in result.iter_mut().skip(1) {
+                *item = Signature::Scalar;
+            }
+            result
+        }
         Function::Aggregate => {
             let mut result = vec![Signature::Vector; arg_count];
             if arg_count >= 1 {
@@ -1526,6 +1533,8 @@ fn static_analysis_on_function(kind: &Function, args: &[Node]) -> StaticResult {
         Function::Drop => not_implemented(args),
         Function::Choosecols => not_implemented(args),
         Function::Chooserows => not_implemented(args),
+        Function::Tocol => not_implemented(args),
+        Function::Torow => not_implemented(args),
         Function::Aggregate => not_implemented(args),
         Function::Rand => not_implemented(args),
         Function::Randbetween => scalar_arguments(args),