+++ crates/ironcalc-base/src/test/mod.rs
@@ -42,6 +42,8 @@ mod test_fn_take;
 mod test_fn_drop;
 mod test_fn_choosecols;
 mod test_fn_chooserows;
+mod test_fn_tocol;
+mod test_fn_torow;
 mod test_fn_transpose;
 mod test_fn_bahttext;
 mod test_fn_percentile;